Partly refactored registration flow
