+++ kano_registration_gui/LabelledEntry.py
@@ -1,10 +1,9 @@
-#!/usr/bin/env python
-
 # LabelledEntry.py
 #
-# Copyright (C) 2015 Kano Computing Ltd.
+# Copyright (C) 2015-2016 Kano Computing Ltd.
 # License: http://www.gnu.org/licenses/gpl-2.0.txt GNU General Public License v2
 #
+
 
 from gi.repository import Gtk, GObject
 from kano_registration_gui.ValidationEntry import ValidationEntry
