+++ kano_registration_gui/TermsAndConditions.py
@@ -25,10 +25,6 @@ class TermsAndConditions(Gtk.Box):
         self.tc_button = OrangeButton(_("I agree to the terms and conditions"))
         self.tc_button.connect("clicked", self._emit_t_and_c_signal)
 
-        # tc_label = self.tc_button.label
-        # tc_label.set_max_width_chars(20)
-        # tc_label.set_line_wrap(True)
-
         self.pack_start(self.checkbutton, False, False, 0)
         self.pack_start(self.tc_button, False, False, 0)
